// src/app/modules/tn-table/interfaces/data-provider.interface.ts
import { Observable } from 'rxjs';
import { TablePagination } from 'app/modules/tn-table/interfaces/table-pagination.interface';
import { TableSort } from 'app/modules/tn-table/interfaces/table-sort.interface';

export interface DataProvider<T> {
  totalRows: number;
  currentPage$: Observable<T[]>;
  currentPageCount$: Observable<number>;
  expandedRow: T | null;
  sorting: TableSort<T>;
  pagination: TablePagination;

  setPagination: (pagination: TablePagination) => void;
  setSorting: (sorting: TableSort<T>) => void;
  setRows: (rows: T[]) => void;
}